fixing @mattgle comments
