+++ App.js
@@ -16,8 +16,7 @@ import { isIos } from '@/constants/platform';
 import Routes from '@/navigation';
 import { initKeyring } from '@/redux/slices/keyring';
 import { persistor, store } from '@/redux/store';
-
-import { TopLevelNavigationRef } from './src/helpers/navigation';
+import { TopLevelNavigationRef } from '@/utils/navigation';
 
 const routingInstrumentation = new Sentry.ReactNavigationInstrumentation();
