+++ src/constants/walletconnect.js
@@ -1,6 +1,10 @@
 export const PLUG_DESCRIPTION = {
   description: 'Plug IC Wallet ',
-  icons: [],
+  icons: [], // We should add urls to icons
   name: 'Plug Wallet',
   url: 'https://plugwallet.ooo',
+};
+
+export const ERRORS = {
+  NO_MSG_SIGN: 'No msg to sign on args',
 };
