+++ src/navigation/navigators/SwipeNavigator.js
@@ -3,14 +3,14 @@ import React, { useEffect } from 'react';
 import { useSelector } from 'react-redux';
 
 import { ENABLE_NFTS } from '@/constants/nfts';
+import useDeepLink from '@/hooks/useDeepLink';
 import NFTs from '@/screens/tabs/NFTs';
 import ProfileScreen from '@/screens/tabs/Profile';
 import Tokens from '@/screens/tabs/Tokens';
+import { useNavigation } from '@/utils/navigation';
 
 import Routes from '../Routes';
-import useDeepLink from '../../hooks/useDeepLink';
 import BottomTabs from './BottomTabs';
-
 
 const Swipe = createMaterialTopTabNavigator();
 
@@ -21,7 +21,9 @@ const SwipeNavigator = () => {
   const navigation = useNavigation();
 
   useEffect(() => {
-    if (deepLink) navigation.navigate(Routes.WALLET_CONNECT);
+    if (deepLink) {
+      navigation.navigate(Routes.WALLET_CONNECT);
+    }
   }, [deepLink]);
 
   useEffect(() => {
@@ -44,7 +46,7 @@ const SwipeNavigator = () => {
       <Swipe.Screen component={ProfileScreen} name={Routes.PROFILE_SCREEN} />
       <Swipe.Screen component={Tokens} name={Routes.TOKENS} />
       {ENABLE_NFTS && <Swipe.Screen component={NFTs} name={Routes.NFTS} />}
-    </Swipe.Navigator >
+    </Swipe.Navigator>
   );
 };
