+++ src/redux/store.js
@@ -100,28 +100,3 @@ export const keyringStorage = {
     ),
   clear: AsyncStorage.clear,
 };
-
-export const walletConnectStorage = {
-  get: async key => {
-    // await AsyncStorage.clear();
-    if (key) {
-      return AsyncStorage.getItem(WALLETCONNECTKEY)
-        .then(value => JSON.parse(`${value}`))
-        .then(parsedValue => parsedValue[key])
-        .then(value => (value ? JSON.parse(value) : value));
-    }
-    return AsyncStorage.getItem(WALLETCONNECTKEY).then(value =>
-      value ? JSON.parse(value) : value,
-    );
-  },
-  set: values =>
-    AsyncStorage.getItem(WALLETCONNECTKEY)
-      .then(savedValues => (savedValues ? JSON.parse(savedValues) : {}))
-      .then(parsedValues =>
-        AsyncStorage.setItem(
-          WALLETCONNECTKEY,
-          Flatted.stringify({ ...parsedValues, values }),
-        ),
-      ),
-  clear: AsyncStorage.setItem(WALLETCONNECTKEY, Flatted.stringify({})),
-};
